:recycle: converter 테스트코드 분리
